Fixed "access arguments" value for WebID Auth config page -- only one permission string.
The default access handler (when the "access_callback" field is undefined), user_access(), accepts only a single permission string. The value previously defined here had a pair of strings (meaning you need both? or either?), but the effect was to pass the second string as the "account" argument, which the function expect to be a User object, not a string. This caused warning messages and prevented access to the configuration page. With the single value, everything works correctly again.